+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - Added an easy way to open replays folder by clicking on the replay saved message, or executing `/guesstheutils replay open`
 - Changed theme length display to individual word lengths
 - Added a configurable offset for current builder's scoreboard line for improved visibility
+- Added options for scoreboard background opacity and highlight strength